New error type
A new error type has been created for the binary part of the tool,
and the project should compile and work. It works with printing out
the errors, and works just like any good old Rust error. There is
however one weak point, and that is debugging is hard with this method.
This should be solved before this is merged into main.