Issue #109 | Improving errors (#184)
* New error type

A new error type has been created for the binary part of the tool,
and the project should compile and work. It works with printing out
the errors, and works just like any good old Rust error. There is
however one weak point, and that is debugging is hard with this method.
This should be solved before this is merged into main.

* Minor edit

* Update src/bin/huak/main.rs

Co-authored-by: Chris Pryer <14341145+cnpryer@users.noreply.github.com>

* Update src/bin/huak/main.rs

Co-authored-by: Chris Pryer <14341145+cnpryer@users.noreply.github.com>

* Changed from CliErrorCode to CliErrorType, prints out anyhow errors now,
and the option for Ruff, PyBlack and PyTest errors can be done. Fixed .gitignore too.

* Update src/huak/errors.rs

Co-authored-by: Chris Pryer <14341145+cnpryer@users.noreply.github.com>

* Fixed main.rs after merging with the updated main branch.

* Formatting

* Making the linter happy

Co-authored-by: Chris Pryer <14341145+cnpryer@users.noreply.github.com>